refactor(test): reframe prod smoke script as the minimal v2 e2e test
Rename scripts/smoke/ -> scripts/minimal_test/ and smoke_v2.py ->
minimal_test_v2.py. Reframe as the minimal "does a v2 batch round-trip
at all?" liveness test rather than a prod smoke test:

- target is now base-URL/key agnostic (SPATIALISE_BASE_URL +
  SPATIALISE_API_KEY) so a dev key against the dev/v2 host is the
  expected first run, per the launch roadmap — no longer prod-locked.
- correct the stale gating note: depends on the v2 host being live
  (V2_DISPATCH_HOST + EE_GET_DATA_QUEUE_NAME set, soilpredict.v2.api
  resolving) + 0.3.0 published, not SPA-1665.
- spell out that this is the minimal liveness check, not the fuller
  test matrix (concurrency, polygon sizes, consistency, edge cases),
  which is tracked separately.

Safety guards (refuse under CI, refuse without explicit config, canary
stub) and the patch-batch read via retrieve_job_status are unchanged.
